修自定义属性判断问题

+++ ConfluenceOtherworld/src/main/java/org/confluence/mod/client/gui/hud/TerraStyleManaHud.java
@@ -48,9 +48,9 @@ public class TerraStyleManaHud implements LayeredDraw.Layer {
             public void render(GuiGraphics guiGraphics, Minecraft minecraft) {
                 int widthMana = guiGraphics.guiWidth() - 21;
                 int heightMana = 4;
-                int currentMana = (int) ClientPacketHandler.getCurrentMana();
+                float currentMana = ClientPacketHandler.getCurrentMana();
                 int maxManaCount = ClientPacketHandler.getMaxMana() / 20;
-                int currentManaToBlit;
+                float currentManaToBlit;
                 float ts;
                 for (int i = 0; i < maxManaCount; i++) {
                     currentManaToBlit = currentMana - (i + 1) * 20;
@@ -58,7 +58,7 @@ public class TerraStyleManaHud implements LayeredDraw.Layer {
                     if (currentManaToBlit >= 0) {
                         guiGraphics.blit(LEGACY_TEXTURE, widthMana + 2, heightMana + i * 12, 18, 34, 13, 16, LEGACY_SIZE, LEGACY_SIZE);
                     } else if (currentManaToBlit + 20 >= 0) {
-                        ts = ((float) (currentManaToBlit + 20)) / 20.0F;
+                        ts = (currentManaToBlit + 20) / 20.0F;
                         guiGraphics.pose().pushPose();
                         guiGraphics.pose().translate(widthMana + 2 + 6.5F * (1 - ts), heightMana + i * 12 + 8.5F * (1 - ts), 0.0F);
                         guiGraphics.pose().scale(ts, ts, 1.0F);
